Level up!
Have finally got Assets to load on to the screen, as well as further
editing the camera to make it work correctly. The original OOP approach
has been altered into an Asset system, where the different cube sizes
inherit from Game Asset (This also includes the diamond) and the
GameAssetManager controls functions - such as drawing the cubes.

This is the next big step in our program, as we've now loaded a
successful level based on the co-ordinates within our .json files.