Improve level.cpp and gameAssetManager
I've created a clear void - emptying the draw_list vector for the next
level to load. Although this may not be necessary - it's an important
exception. I've also blackened the screen once the level is completed,
now I just need to find some text to say "Level completed! Loading... ",
but until then there is a black screen.