Mouse Motion Event Fix
Had an issue where the camera would look around even when no one was
moving the mouse. The issue was that there was nothing to tell the mouse
motion event when to stop computing the mouse angles, so it kept moving.
I have put in a line of code to ignore the event once it has finished
computing.