The Single Asset
I have created a loop which now increases the small cubes size, into a
medium or large cube. This meant that I no longer needed individual
assets for each of them, and could replace one of them with a door
asset.

Also I have coded the logic required for collision, and now everything
has been setup for its completion. All that's left now is to get the
logic working and then it will all be working correctly.