Thou Shall Now Touch!
Collision is up and running, allowing for level transitions, diamond
collecting and gravity to occur. There are, however, issues with cube
collision which I intend to work on through the next couple of days,
although level transitioning is now possible!

Other changes include a position reset for the camera class, meaning the
player will now return to the starting position at the beginning of
every new level. Level had an issue where an int jumped from 1 to
4203678 - for reasons I'm currently unsure of - so I changed a method to
return a solid number and, in the process, fix the problem.