Untested Change
As mentioned I was having problems with cube collision, and I'm hoping
that this will help fix the problem. Once I get round to testing this on
my other machine, I will be able to come to a conclusion on how to fix
the issue.

Previously the for loop found if the player "wasn't" colliding with any
cubes, and then returned that gravity is true. This way is reversed and
returns false if the player IS colliding with any of the cubes. This way
should cause the loop to go through EVERY cube, rather than stopping on
the first cube its not colliding with.