Experimental Change
I hope to have made a start, or even finished, the implementation of a
"shy" intelligence to the diamonds. This will cause them to retreat in
the opposite direction of the player, once they invade a space around
the diamond, so long as the diamonds are colliding with the boxes.

This change is, as stated, experimental and may or may not be removed
after extensive testing.