"Today is a good day"
Previously we were having major issues with crashing, regarding one of
our vectors, at random points during any of our levels. After searching
through the program I learned that our int "cubes" was not getting reset
after every level, and after resetting it the problem was resolved.
*Sigh of relief*
Another change is that the player now needs to collect all the diamonds,
on each level, before they can progress through the door and on to the
next level. This means that the Platformer now has an objective which
players will need to complete. This will also mean adding more diamonds
to levels in the future.
Finally the testing of the diamond intelligence is complete and,
although there are still small issues that need to be addressed, they
now flee in the opposite direction of the player - if they invade the
space around them. Oh and diamonds are now collectable and are removed
from the level when collected!