Changes galore!
Last day to commit our work so we made sure everything was in order.
Doxygen now has a Main page, bug list (Due to expand) and a lot more
comments on level.cpp. We've also, as a goodbye to our program, added a
day-to-night system which occurs during the game! Several of our .cpps
have been edited to make the game use less code and variables, or have
had changes to make the game feel more interactive.